Extract tests into dedicated modules
