+++ streamsockets-auth/Cargo.toml
@@ -26,6 +26,7 @@ blake3 = { version = "1.5", default-features = false }
 [dev-dependencies]
 tokio = { workspace = true, features = ["full"] }
 tempfile = { workspace = true }
+tracing-subscriber = { workspace = true }
 
 [lints]
 workspace = true
