+++ streamsockets-auth/src/lib.rs
@@ -55,7 +55,7 @@ use std::net::IpAddr;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
 use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Arc;
+use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
 use arc_swap::ArcSwap;
@@ -529,6 +529,7 @@ pub struct TokenAuthentication {
     snapshot: Arc<ArcSwap<AccountsSnapshot>>,
     leases: Arc<LeaseTracker>,
     config_path: Option<PathBuf>,
+    config_hash: Mutex<Option<[u8; 32]>>,
     reload_lock: tokio::sync::Mutex<()>,
 }
 
@@ -537,12 +538,13 @@ impl TokenAuthentication {
     /// failures preserve the original snapshot.
     pub fn from_file(path: impl AsRef<Path>) -> Result<Self, AuthError> {
         let path = path.as_ref().to_path_buf();
-        let file = load_yaml_blocking(&path)?;
-        let snapshot = Arc::new(ArcSwap::from_pointee(AccountsSnapshot::build(file)?));
+        let loaded = load_yaml_blocking(&path)?;
+        let snapshot = Arc::new(ArcSwap::from_pointee(AccountsSnapshot::build(loaded.file)?));
         Ok(Self {
             snapshot,
             leases: Arc::new(LeaseTracker::new()),
             config_path: Some(path),
+            config_hash: Mutex::new(Some(loaded.digest)),
             reload_lock: tokio::sync::Mutex::new(()),
         })
     }
@@ -554,6 +556,7 @@ impl TokenAuthentication {
             snapshot,
             leases: Arc::new(LeaseTracker::new()),
             config_path: None,
+            config_hash: Mutex::new(None),
             reload_lock: tokio::sync::Mutex::new(()),
         })
     }
@@ -585,12 +588,24 @@ impl TokenAuthentication {
             return;
         };
         let _g = self.reload_lock.lock().await;
-        match load_yaml_async(path)
-            .await
-            .and_then(AccountsSnapshot::build)
-        {
+        let loaded = match load_yaml_async(path).await {
+            Ok(loaded) => loaded,
+            Err(e) => {
+                warn!(path = %path.display(), error = %e, "reload failed; keeping previous snapshot");
+                return;
+            }
+        };
+        let mut config_hash = self
+            .config_hash
+            .lock()
+            .unwrap_or_else(|e| e.into_inner());
+        if config_hash.as_ref() == Some(&loaded.digest) {
+            return;
+        }
+        match AccountsSnapshot::build(loaded.file) {
             Ok(new_snap) => {
                 self.snapshot.store(Arc::new(new_snap));
+                *config_hash = Some(loaded.digest);
                 info!(path = %path.display(), "reloaded accounts");
             }
             Err(e) => {
@@ -657,7 +672,17 @@ impl TokenAuthentication {
     }
 }
 
-fn load_yaml_blocking(path: &Path) -> Result<AccountsFile, AuthError> {
+#[derive(Debug)]
+struct LoadedAccounts {
+    file: AccountsFile,
+    digest: [u8; 32],
+}
+
+fn digest_bytes(bytes: &[u8]) -> [u8; 32] {
+    *blake3::hash(bytes).as_bytes()
+}
+
+fn load_yaml_blocking(path: &Path) -> Result<LoadedAccounts, AuthError> {
     let meta = std::fs::metadata(path).map_err(|source| AuthError::Io {
         path: path.to_path_buf(),
         source,
@@ -673,13 +698,15 @@ fn load_yaml_blocking(path: &Path) -> Result<AccountsFile, AuthError> {
         path: path.to_path_buf(),
         source,
     })?;
-    serde_yml::from_slice(&bytes).map_err(|e| AuthError::Parse {
+    let digest = digest_bytes(&bytes);
+    let file = serde_yml::from_slice(&bytes).map_err(|e| AuthError::Parse {
         path: path.to_path_buf(),
         location: redact_yaml_location(&e),
-    })
+    })?;
+    Ok(LoadedAccounts { file, digest })
 }
 
-async fn load_yaml_async(path: &Path) -> Result<AccountsFile, AuthError> {
+async fn load_yaml_async(path: &Path) -> Result<LoadedAccounts, AuthError> {
     let meta = tokio::fs::metadata(path)
         .await
         .map_err(|source| AuthError::Io {
@@ -699,10 +726,12 @@ async fn load_yaml_async(path: &Path) -> Result<AccountsFile, AuthError> {
             path: path.to_path_buf(),
             source,
         })?;
-    serde_yml::from_slice(&bytes).map_err(|e| AuthError::Parse {
+    let digest = digest_bytes(&bytes);
+    let file = serde_yml::from_slice(&bytes).map_err(|e| AuthError::Parse {
         path: path.to_path_buf(),
         location: redact_yaml_location(&e),
-    })
+    })?;
+    Ok(LoadedAccounts { file, digest })
 }
 
 const INOTIFY_QUIESCE: Duration = Duration::from_millis(100);
@@ -801,310 +830,4 @@ async fn run_inotify(path: PathBuf, auth: Arc<TokenAuthentication>) -> Result<()
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn sample_file() -> AccountsFile {
-        AccountsFile {
-            accounts: vec![
-                Account {
-                    name: "user1".into(),
-                    token: "tok1".into(),
-                    reuse: false,
-                    routes: vec!["127.0.0.1:8888".into()],
-                    allowed_ips: vec!["127.0.0.1".into(), "192.168.1.0/24".into()],
-                },
-                Account {
-                    name: "user2".into(),
-                    token: "tok2".into(),
-                    reuse: true,
-                    routes: vec!["example.com:5050".into()],
-                    allowed_ips: vec!["10.0.0.0/8".into()],
-                },
-            ],
-        }
-    }
-
-    fn ip(s: &str) -> IpAddr {
-        s.parse().expect("test")
-    }
-
-    #[test]
-    fn build_and_query() {
-        let snap = AccountsSnapshot::build(sample_file()).expect("test");
-        let m = snap.authenticate("tok1", "127.0.0.1:8888", ip("127.0.0.1"));
-        assert!(m.is_some());
-        assert_eq!(m.expect("test").account.name, "user1");
-        assert!(snap
-            .authenticate("tok1", "127.0.0.1:8888", ip("192.168.1.50"))
-            .is_some());
-        assert!(snap
-            .authenticate("tok1", "127.0.0.1:8888", ip("10.0.0.1"))
-            .is_none());
-        assert!(snap
-            .authenticate("badtoken", "127.0.0.1:8888", ip("127.0.0.1"))
-            .is_none());
-        assert!(snap
-            .authenticate("tok1", "wrong.route:1", ip("127.0.0.1"))
-            .is_none());
-    }
-
-    #[test]
-    fn duplicate_token_rejected_reports_account_name_not_token() {
-        let mut f = sample_file();
-        f.accounts[1].token = "tok1".into();
-        let r = AccountsSnapshot::build(f);
-        match r {
-            Err(AuthError::DuplicateToken(name)) => {
-                assert_eq!(name, "user2");
-                let displayed = format!("{}", AuthError::DuplicateToken(name));
-                assert!(!displayed.contains("tok1"));
-            }
-            other => panic!("expected DuplicateToken, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn lease_no_reuse_conflicts() {
-        let auth = TokenAuthentication::from_accounts(sample_file()).expect("test");
-        let snap = auth.snapshot_arc();
-        let hash = token_hash(b"tok1");
-        let acc = snap.by_token_hash.get(&hash).expect("test").account.clone();
-        let g1 = auth.leases().try_lease_arc(&acc).expect("test");
-        let r2 = auth.leases().try_lease_arc(&acc);
-        assert!(matches!(r2, Err(LeaseError::Conflict)));
-        drop(g1);
-        let _g3 = auth.leases().try_lease_arc(&acc).expect("test");
-    }
-
-    #[test]
-    fn lease_reuse_true_allows_concurrent() {
-        let auth = TokenAuthentication::from_accounts(sample_file()).expect("test");
-        let snap = auth.snapshot_arc();
-        let hash = token_hash(b"tok2");
-        let acc = snap.by_token_hash.get(&hash).expect("test").account.clone();
-        let g1 = auth.leases().try_lease_arc(&acc).expect("test");
-        let g2 = auth.leases().try_lease_arc(&acc).expect("test");
-        assert_eq!(auth.leases().active_count(&acc), 2);
-        drop(g1);
-        drop(g2);
-        assert_eq!(auth.leases().active_count(&acc), 0);
-    }
-
-    #[test]
-    fn lease_cas_serializes_concurrent_attempts() {
-        use std::sync::atomic::AtomicUsize;
-        let auth = Arc::new(TokenAuthentication::from_accounts(sample_file()).expect("test"));
-        let snap = auth.snapshot_arc();
-        let hash = token_hash(b"tok1");
-        let acc = snap.by_token_hash.get(&hash).expect("test").account.clone();
-
-        let won = Arc::new(AtomicUsize::new(0));
-        let conflict = Arc::new(AtomicUsize::new(0));
-        let mut handles = Vec::new();
-        for _ in 0..16 {
-            let leases = auth.leases().clone();
-            let acc = acc.clone();
-            let won = won.clone();
-            let conflict = conflict.clone();
-            handles.push(std::thread::spawn(move || {
-                match leases.try_lease_arc(&acc) {
-                    Ok(g) => {
-                        won.fetch_add(1, Ordering::SeqCst);
-                        drop(g);
-                    }
-                    Err(LeaseError::Conflict) => {
-                        conflict.fetch_add(1, Ordering::SeqCst);
-                    }
-                    Err(LeaseError::Exhausted) => {
-                        panic!("unexpected Exhausted on reuse=false path");
-                    }
-                }
-            }));
-        }
-        for h in handles {
-            h.join().expect("test");
-        }
-        assert_eq!(
-            won.load(Ordering::SeqCst) + conflict.load(Ordering::SeqCst),
-            16
-        );
-        assert_eq!(auth.leases().active_count(&acc), 0);
-    }
-
-    #[test]
-    fn lease_per_account_ceiling_enforced() {
-        let leases = Arc::new(LeaseTracker::with_max_per_account(3));
-        let acc = Arc::new(Account {
-            name: "u".into(),
-            token: "t".into(),
-            reuse: true,
-            routes: vec![],
-            allowed_ips: vec![],
-        });
-        let g1 = leases.try_lease_arc(&acc).expect("test");
-        let _g2 = leases.try_lease_arc(&acc).expect("test");
-        let _g3 = leases.try_lease_arc(&acc).expect("test");
-        let r = leases.try_lease_arc(&acc);
-        assert!(matches!(r, Err(LeaseError::Exhausted)));
-        assert_eq!(leases.active_count(&acc), 3);
-        drop(g1);
-        let _g4 = leases.try_lease_arc(&acc).expect("test");
-        assert_eq!(leases.active_count(&acc), 3);
-    }
-
-    #[test]
-    fn lease_release_garbage_collects_zero_entries() {
-        let leases = Arc::new(LeaseTracker::new());
-        let acc = Arc::new(Account {
-            name: "u".into(),
-            token: "tok-gc".into(),
-            reuse: true,
-            routes: vec![],
-            allowed_ips: vec![],
-        });
-        let g = leases.try_lease_arc(&acc).expect("test");
-        let h = token_hash(b"tok-gc");
-        assert!(leases.active.contains_key(&h));
-        drop(g);
-        // After release, the entry should be GCed so reloads cannot leak.
-        assert!(!leases.active.contains_key(&h));
-    }
-
-    #[test]
-    fn lease_survives_reload_when_token_unchanged() {
-        // Two snapshots issuing fresh Arc<Account> with the same token must
-        // resolve to the same lease slot.
-        let snap1 = AccountsSnapshot::build(sample_file()).expect("test");
-        let snap2 = AccountsSnapshot::build(sample_file()).expect("test");
-        let h = token_hash(b"tok2");
-        let acc1 = snap1.by_token_hash.get(&h).expect("test").account.clone();
-        let acc2 = snap2.by_token_hash.get(&h).expect("test").account.clone();
-        // They are distinct Arc allocations:
-        assert!(!Arc::ptr_eq(&acc1, &acc2));
-
-        let leases = Arc::new(LeaseTracker::new());
-        let g1 = leases.try_lease_arc(&acc1).expect("test");
-        // Release-via-acc2 path: hash-keyed tracker must observe count==1
-        // for either Arc identity.
-        assert_eq!(leases.active_count(&acc1), 1);
-        assert_eq!(leases.active_count(&acc2), 1);
-        let g2 = leases.try_lease_arc(&acc2).expect("test");
-        assert_eq!(leases.active_count(&acc1), 2);
-        drop(g1);
-        drop(g2);
-        assert_eq!(leases.active_count(&acc1), 0);
-    }
-
-    /// Smoke test for timing-attack invariance on the HashMap probe.
-    ///
-    /// We don't claim ns-tight constant time — that would require disabling
-    /// CPU frequency scaling, isolating cores, and a microbenchmarking
-    /// harness. What we *do* claim: a random 32-byte input and a known-prefix
-    /// input both hit the same code path (BLAKE3 hash → HashMap miss →
-    /// return None) with no token-content-dependent branches before the miss.
-    /// This test is a sanity check that the order-of-magnitude is comparable.
-    #[test]
-    fn authenticate_timing_smoke() {
-        use std::time::Instant;
-        let snap = AccountsSnapshot::build(sample_file()).expect("test");
-        let ip4 = ip("127.0.0.1");
-
-        // Warm up.
-        for _ in 0..10_000 {
-            let _ = snap.authenticate("tok-no-match", "r", ip4);
-            let _ = snap.authenticate("tok1prefix-extended-bytes-x", "r", ip4);
-        }
-
-        let n = 50_000;
-        let t0 = Instant::now();
-        for _ in 0..n {
-            // Token close to a real one: same length, different content.
-            let _ = snap.authenticate("tok9", "r", ip4);
-        }
-        let near_miss = t0.elapsed();
-
-        let t1 = Instant::now();
-        for _ in 0..n {
-            // Token nothing like a real one.
-            let _ = snap.authenticate("zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz", "r", ip4);
-        }
-        let random_miss = t1.elapsed();
-
-        // Tolerate a 5x ratio either direction. Lower bounds aren't meaningful
-        // here — we're guarding against an order-of-magnitude regression
-        // (e.g. accidentally adding a strncmp short-circuit before the hash).
-        let ratio = near_miss.as_nanos() as f64 / random_miss.as_nanos() as f64;
-        assert!(
-            (0.2..=5.0).contains(&ratio),
-            "near_miss={near_miss:?} random_miss={random_miss:?} ratio={ratio}"
-        );
-    }
-
-    #[tokio::test]
-    async fn reload_swap_and_malformed_preserves_snapshot() {
-        use std::io::Write;
-        let mut tf = tempfile::NamedTempFile::new().expect("test");
-        let yaml1 = r#"
-accounts:
-  - name: u1
-    token: tok-old
-    reuse: true
-    routes: ["127.0.0.1:1"]
-    allowedIps: ["127.0.0.1"]
-"#;
-        tf.write_all(yaml1.as_bytes()).expect("test");
-        tf.flush().expect("test");
-
-        let auth = TokenAuthentication::from_file(tf.path()).expect("test");
-        let snap1 = auth.snapshot_arc();
-        let h_old = token_hash(b"tok-old");
-        let h_new = token_hash(b"tok-new");
-        assert!(snap1.by_token_hash.contains_key(&h_old));
-        assert!(!snap1.by_token_hash.contains_key(&h_new));
-
-        let yaml2 = r#"
-accounts:
-  - name: u1
-    token: tok-new
-    reuse: true
-    routes: ["127.0.0.1:1"]
-    allowedIps: ["127.0.0.1"]
-"#;
-        std::fs::write(tf.path(), yaml2).expect("test");
-        auth.reload().await;
-        let snap2 = auth.snapshot_arc();
-        assert!(snap2.by_token_hash.contains_key(&h_new));
-        assert!(!snap2.by_token_hash.contains_key(&h_old));
-
-        std::fs::write(tf.path(), b": : : not yaml").expect("test");
-        auth.reload().await;
-        let snap3 = auth.snapshot_arc();
-        assert!(
-            snap3.by_token_hash.contains_key(&h_new),
-            "malformed reload must preserve previous snapshot"
-        );
-
-        let yaml_dup = r#"
-accounts:
-  - name: u1
-    token: dup
-    reuse: true
-    routes: ["127.0.0.1:1"]
-    allowedIps: ["127.0.0.1"]
-  - name: u2
-    token: dup
-    reuse: true
-    routes: ["127.0.0.1:1"]
-    allowedIps: ["127.0.0.1"]
-"#;
-        std::fs::write(tf.path(), yaml_dup).expect("test");
-        auth.reload().await;
-        let snap4 = auth.snapshot_arc();
-        assert!(
-            snap4.by_token_hash.contains_key(&h_new),
-            "duplicate-token reload must preserve previous snapshot"
-        );
-        assert!(!snap4.by_token_hash.contains_key(&token_hash(b"dup")));
-    }
-}
+mod tests;
