+++ streamsockets-core/src/tests/mod.rs
@@ -1,0 +1,55 @@
+use super::*;
+
+#[test]
+fn env_value_default() {
+    let v = env_value("__STREAMSOCKETS_TEST_DEFAULT__", "fallback");
+    assert_eq!(v, "fallback");
+}
+
+#[test]
+fn env_int_default() {
+    let v = env_value_as_int("__STREAMSOCKETS_TEST_INT__", 42);
+    assert_eq!(v, 42);
+}
+
+#[test]
+fn env_bool_parses_variants() {
+    std::env::set_var("__BOOL_T__", "TRUE");
+    assert!(env_bool("__BOOL_T__", false));
+    std::env::set_var("__BOOL_F__", "no");
+    assert!(!env_bool("__BOOL_F__", true));
+    std::env::remove_var("__BOOL_T__");
+    std::env::remove_var("__BOOL_F__");
+}
+
+/// Non-Linux dev builds always pick Tokio: the runtime selector short-
+/// circuits before any kernel probe.
+#[cfg(not(target_os = "linux"))]
+#[test]
+fn pick_runtime_non_linux_is_tokio() {
+    assert_eq!(pick_runtime(), RuntimeKind::Tokio);
+    assert_eq!(RuntimeKind::Tokio.label(), "tokio");
+}
+
+/// Linux: with `DISABLE_IOURING=true` the selector returns `Tokio` and the
+/// label is `epoll` (kernel-side mechanism under tokio's reactor).
+#[cfg(target_os = "linux")]
+#[test]
+fn pick_runtime_disable_iouring_is_epoll() {
+    // The DISABLE_IOURING decision is cached by `is_iouring_disabled()`
+    // on first call (process-wide OnceCell). To make this test
+    // deterministic we set the env var before any test that reads it
+    // runs — but tests run in parallel and cargo test won't fork. We
+    // therefore exercise pick_runtime's *fallback path* indirectly by
+    // asserting the label invariant: whatever pick_runtime returns, its
+    // label must be one of the documented stable values.
+    let rt = pick_runtime();
+    let label = rt.label();
+    assert!(
+        matches!(label, "io_uring" | "epoll"),
+        "unexpected label on Linux: {label}"
+    );
+    // RuntimeKind::Tokio.label() on Linux is "epoll" by spec.
+    assert_eq!(RuntimeKind::Tokio.label(), "epoll");
+    assert_eq!(RuntimeKind::Uring.label(), "io_uring");
+}
